requirements.txt
streamlit
google-genai
pymongo
